udev: lock symlink directories in link_update()
When uevents are processed in parallel, in particular during coldplug,
it can happen that several events for the same symlink arrive at the
same time (e.g. SCSI disk and for a multipath DM device for that
disk). This opens up a race window where a higher-priority link target
may not be visible under /run/udev/links/ yet when a lower-priority
one is processed, leading to the creation of the wrong symlink.

The recently introduced "multiple symlinks" test from the udev test suite
reveals this race condition, most easily if run on a system with many cores.

We have to avoid this by locking the directory under /run/udev/links/.
I have extensively studied alternative approaches; nothing besides
locking avoids the said race condition reliably.

Initial attempts to do this with systemd's make_lock_file_for() API
revealed another problem: In certain situations, there may be a lot of
contenders for a given symlink (for example, on various distributions, almost
all GPT partitions have the partition labels "primary" or "logical" due
to a bug in past parted releases - I have seen systems with literally
thousands of contenders for the /dev/disk/by-partlabel/primary symlink).
As of 4.18, the kernel implementation of fcntl(F_OFD_SETLKW) is inefficient,
leading to busy-waiting in spin-locks and thus very bad scaling when
many contending processes compete for a file lock. The kernel implementation
can be improved and patches are under discussion, but it will take a long
time until that kernel-side improvement is generally available.

To avoid this problem, this patch doesn't use make_lock_file_for().
Rather, it uses a set of SysV semaphores, where the index into the semaphore
set is derived from the symlink name using udev's built-in hash function.
This yields good collision avoidance for different symlink names and performs
way better than fcntl(F_OFD_SETLKW) with many contenders.

+++ src/udev/udev-node.c
@@ -8,6 +8,7 @@
 #include <string.h>
 #include <sys/stat.h>
 #include <unistd.h>
+#include <sys/sem.h>
 
 #include "alloc-util.h"
 #include "device-nodes.h"
@@ -26,6 +27,9 @@
 #include "string-util.h"
 #include "strxcpyx.h"
 #include "udev-node.h"
+#include "siphash24.h"
+#include "hash-funcs.h"
+#include "udev.h"
 
 static int node_symlink(sd_device *dev, const char *node, const char *slink) {
         _cleanup_free_ char *slink_dirname = NULL, *target = NULL;
@@ -185,15 +189,102 @@ static int link_find_prioritized(sd_device *dev, bool add, const char *stackdir,
         return 0;
 }
 
+#define NSEM 1024
+static int init_link_semaphores(const char *path)
+{
+        key_t key = ftok(path, 0);
+        int semid;
+
+        semid = semget(key, NSEM, 0600|IPC_CREAT|IPC_EXCL);
+        if (semid != -1) {
+                unsigned short val[NSEM];
+                int i;
+                struct sembuf dummy_op[]  = {
+                        { .sem_num = 0, .sem_op = -1, .sem_flg = 0 },
+                        { .sem_num = 0, .sem_op = 1, .sem_flg = 0 },
+                };
+
+                for (i = 0; i < NSEM; i++)
+                        val[i] = 1;
+                if (semctl(semid, 0, SETALL, val) == -1)
+                        return -errno;
+
+                /* Dummy semop to set sem_otime */
+                if (semop(semid, dummy_op, (sizeof(dummy_op)/sizeof(*dummy_op)))
+                    == -1)
+                        return -errno;
+                return semid;
+        } else {
+                const int RETRIES = 10;
+                int i;
+
+                semid = semget(key, NSEM, 0);
+                if (semid == -1)
+                        return -errno;
+
+                for (i = 0; i < RETRIES; i++) {
+                        struct semid_ds ds;
+
+                        /* Wait for initialization to finish */
+                        if (semctl(semid, 0, IPC_STAT, &ds) != -1 &&
+                            ds.sem_otime != 0)
+                                return semid;
+                        usleep(10000);
+                }
+                return -1;
+        }
+}
+
+static unsigned short get_sema_index(const char *link)
+{
+        static const unsigned char seed[16] = { 0x6b, 0xb0, 0xb1, 0x28,
+                                                0xf7, 0x8c, 0x59, 0xb2,
+                                                0x05, 0x1d, 0xd1, 0xa2,
+                                                0xcc, 0x12, 0xae, 0xb7 };
+        struct siphash state;
+        uint64_t hash;
+
+        siphash24_init(&state, seed);
+        path_hash_func(link, &state);
+        hash = siphash24_finalize(&state);
+
+        return hash & (NSEM-1);
+}
+
+static int _slink_semop(int semid, unsigned short semidx,
+                        int op, const char *msg)
+{
+        struct sembuf sb = { .sem_num = semidx, .sem_op = op, .sem_flg = 0 };
+
+        if (semop(semid, &sb, 1) == -1)
+                return log_warning_errno(-errno, "failed to %s semaphore", msg);
+        return 0;
+}
+
+#define lock_slink(semid, semidx) \
+        _slink_semop((semid), (semidx), -1, "acquire")
+#define unlock_slink(semid, semidx) \
+        _slink_semop((semid), (semidx), 1, "release")
+
 /* manage "stack of names" with possibly specified device priorities */
 static int link_update(sd_device *dev, const char *slink, bool add) {
         _cleanup_free_ char *target = NULL, *filename = NULL, *dirname = NULL;
         char name_enc[PATH_MAX];
         const char *id_filename;
+        static int semid = -1;
+        unsigned short semidx;
         int r;
 
         assert(dev);
         assert(slink);
+
+        if (semid == -1) {
+                semid = init_link_semaphores("/run/udev/links");
+                if (semid == -1)
+                        return log_device_error_errno(dev, r,
+                                                      "failed to set up semaphores: %m");
+        }
+        semidx = get_sema_index(slink);
 
         r = device_get_id_filename(dev, &id_filename);
         if (r < 0)
@@ -206,6 +297,11 @@ static int link_update(sd_device *dev, const char *slink, bool add) {
         filename = path_join(dirname, id_filename);
         if (!filename)
                 return log_oom();
+
+        mkdir_parents(dirname, 0755);
+        r = lock_slink(semid, semidx);
+        if (r != 0)
+                return r;
 
         if (!add && unlink(filename) == 0)
                 (void) rmdir(dirname);
@@ -230,6 +326,7 @@ static int link_update(sd_device *dev, const char *slink, bool add) {
                                 r = -errno;
                 } while (r == -ENOENT);
 
+        (void)unlock_slink(semid, semidx);
         return r;
 }
